Added DAO

+++ src/main/java/pl/app/building/equipment/model/Customer.java
@@ -21,7 +21,7 @@ public class Customer {
     private int id;
     private String name;
     private String surname;
-    private int pesel;
+    private long pesel;
     private int documentNumber;
 
     @ManyToMany(mappedBy = "customers")
